requirements.txt
streamlit>=1.33.0
groq>=0.11.0
python-dotenv>=1.0.0
